+++ packages/create-lib/src/AppGenerator/AppGenerator.ts
@@ -4,12 +4,12 @@ import { join, basename } from 'path';
 export default class AppGenerator extends Generator {
   async writing() {
     const cwd = winPath(this.cwd);
-    const packageName = this.args.name || basename(cwd);
+    const packageName =
+      this.args.name || this.args.packageName || basename(cwd);
 
     return await this.copyDirectory({
       context: {
-        version: require('../../package').version,
-        siteMode: this.args.site,
+        ...this.args,
         packageName,
       },
       path: join(__dirname, '../../tpls/AppGenerator'),